use-sort-length-every-lowerCase-indexOf-split-to-solve
First of, use sort to change the string into an array. Then subtract one from the other and if they contain similar letters, then return true. Use every to check through the lettering. Include lowerCase to standardise the different strings. Use indexOf to return a specified letter in a string.