use-str-arr-regex-ascii-boolean-ifelse-to-solve
I had to turn a #string into an #array to manipulate it, then turned it back into a string. Then I used #regex to deal with non alphabetical characters. After this I used a #boolean to decide if true or false. Then I used  an #ascii code to move letters. Finally, rounded up with a #ifelse statement to resolve it.