use-split-regex-join-while-&-if-statements-to-solve
Create an array of characters from the original string, Pull out the vowels and test the first character of the array. If there are matches then add way onto the end of the string. Where there are no matching vowels, push it to the end and splice the !st letter. If a vowel is found then it breaks Put ay on the end to bring the string back together.