Use-array-push-forloop-ifstatements-to-solve
Create a duplicate array from the original array. Iterate twice through the length of the arguments in the forloop. Check if the placehoder contains the value of arguments. Return -1 if it doesn't. Use push to add it to the end of the array. Then set the original array to the altered array.