use-regex-and-replace-methods-to-solve
Convert string so that every word is lower case and it's joined by dashes. They are known as spinal case. Change the string when it has underscores and white space characters. Also change string to lower case in order to return it. Use a capturing group in regex to take care of replace method.