use-array-while-loop-filter-push-length-methods-to-solve
I created a global array so that all code can sit under it. Start with the number 1 because this is the beginning of the fibonacci sequence. Use the while loop to decide if the numbers are less or equal to the number argument. If so, pass them and push onto the end of the code sequence. Lastly, filter out and reduce the odd numbers.