Solved-by-first-creating-an-array-prototype
Logic:
Create new array property identifying the last value.
Create new array property identifying the second to last value.
Create array starting the fibonacci sequence.
Continue the sequence (by adding the last 2 numbers in the array) while the next fib number is < (num)
Remove all even numbers from the array (filter).
Add all numbers in array (reduce).

NB: Array prototypes allow you to add in your own methods. Therefore, you can take in a function and return, this. This relates to the prototype that's been called on, i.e. if an array was a bunch of numbers, then this, relates to it!